GeoBoundary implemented
GeoBoundary is switched off by default as every movement of the map results in reloading data.
If the amount of loaded features gets too high, it is recommended to switch it on.
Some minor tweaks to CSS to show less distorted Pokéstop images.
Added settings for pitch and bearing to save rotated or angled views.